Refactor into seperable compilation units
Build as single extension module

Make implementation module private (_numcertain)
