+++ setup.py
@@ -3,16 +3,19 @@ from sysconfig import get_paths
 from numpy import get_include
 from numpy.distutils.core import Extension, setup
 
-ext_modules = [
+INCLUDE_DIRS = [get_paths()["include"], get_include()]
+
+EXT_MODULES = [
     Extension(
-        "numcertain.uncertain",
-        [
-            "src/numcertain/uncertain.c",
-            "src/numcertain/uncertaindtype.c",
-            "src/numcertain/npyuncertain.c",
+        name="numcertain._numcertain",
+        sources=[
+            "src/numcertain/numcertain.c",
+            "src/numcertain/numpy.c",
+            "src/numcertain/pytype.c",
+            "src/numcertain/ctype.c",
         ],
-        include_dirs=[get_paths()["include"], get_include()],
+        include_dirs=INCLUDE_DIRS,
     ),
 ]
 
-setup(ext_modules=ext_modules)
+setup(ext_modules=EXT_MODULES)
