+++ src/numcertain/numcertain.h
@@ -1,0 +1,6 @@
+#pragma once
+
+#include "numpy.h"
+#include "pytype.h"
+
+PyMODINIT_FUNC PyInit__numcertain(void);
