+++ src/numcertain/numpy.c
@@ -1,4 +1,4 @@
-#include "npyuncertain.h"
+#include "numpy.h"
 
 PyObject *npyuncertain_getitem(void *data, void *arr) {
   Uncertain_t u;
@@ -12,8 +12,6 @@ int npyuncertain_setitem(PyObject *item, void *data, void *arr) {
     u = ((PyUncertain_t *)item)->u;
   else {
     long long n = PyLong_AsLongLong(item);
-    if (error_converting(n))
-      return -1;
     PyObject *y = PyLong_FromLongLong(n);
     if (!y)
       return -1;
@@ -174,3 +172,144 @@ UNCERTAIN_BINOP_UFUNC(not_equal, bool, uncertain_ne)
 
 UNCERTAIN_UNARY_UFUNC(nominal, npy_float64, uncertain_nominal)
 UNCERTAIN_UNARY_UFUNC(uncertainty, npy_float64, uncertain_uncertainty)
+
+static PyMethodDef module_methods[] = {{0}};
+
+static struct PyModuleDef moduledef = {PyModuleDef_HEAD_INIT,
+                                       "numpy",
+                                       NULL,
+                                       -1,
+                                       module_methods,
+                                       NULL,
+                                       NULL,
+                                       NULL,
+                                       NULL};
+
+int add_numpy_functionality(PyObject *module) {
+
+  import_array();
+  import_umath();
+
+  PyObject *numpy_str = PyUnicode_FromString("numpy");
+  if (!numpy_str)
+    return -1;
+  PyObject *numpy = PyImport_Import(numpy_str);
+  Py_DECREF(numpy_str);
+  if (!numpy)
+    return -1;
+
+  PyUncertain_Type.tp_base = &PyGenericArrType_Type;
+
+  if (PyType_Ready(&PyUncertain_Type) < 0)
+    return -1;
+
+  PyArray_InitArrFuncs(&npyuncertain_arrfuncs);
+  npyuncertain_arrfuncs.getitem = npyuncertain_getitem;
+  npyuncertain_arrfuncs.setitem = npyuncertain_setitem;
+  npyuncertain_arrfuncs.copyswapn = npyuncertain_copyswapn;
+  npyuncertain_arrfuncs.copyswap = npyuncertain_copyswap;
+  npyuncertain_arrfuncs.nonzero = npyuncertain_nonzero;
+
+  Py_SET_TYPE(&npyuncertain_descr, &PyArrayDescr_Type);
+
+  int npy_uncertain = PyArray_RegisterDataType(&npyuncertain_descr);
+  if (npy_uncertain < 0) {
+    return -1;
+  }
+
+  if (PyDict_SetItemString(PyUncertain_Type.tp_dict, "dtype",
+                           (PyObject *)&npyuncertain_descr) < 0)
+    return -1;
+
+#define REGISTER_CAST(From, To, from_descr, to_typenum, safe)                  \
+  {                                                                            \
+    if (PyArray_RegisterCastFunc((from_descr), (to_typenum),                   \
+                                 npycast_##From##_##To) < 0)                   \
+      return -1;                                                               \
+    if (safe &&                                                                \
+        PyArray_RegisterCanCast((from_descr), (to_typenum), NPY_NOSCALAR) < 0) \
+      return -1;                                                               \
+  };
+
+#define REGISTER_FLOAT_CAST(bits)                                              \
+  REGISTER_CAST(npy_float##bits, Uncertain_t,                                  \
+                PyArray_DescrFromType(NPY_FLOAT##bits), npy_uncertain, 1)      \
+  REGISTER_CAST(Uncertain_t, npy_float##bits, &npyuncertain_descr,             \
+                NPY_FLOAT##bits, 0)
+
+#define REGISTER_INT_CAST(bits)                                                \
+  REGISTER_CAST(npy_int##bits, Uncertain_t,                                    \
+                PyArray_DescrFromType(NPY_INT##bits), npy_uncertain, 1)        \
+  REGISTER_CAST(Uncertain_t, npy_int##bits, &npyuncertain_descr,               \
+                NPY_INT##bits, 0)
+
+  REGISTER_FLOAT_CAST(32)
+  REGISTER_FLOAT_CAST(64)
+
+  REGISTER_INT_CAST(8)
+  REGISTER_INT_CAST(16)
+  REGISTER_INT_CAST(32)
+  REGISTER_INT_CAST(64)
+
+  REGISTER_CAST(npy_bool, Uncertain_t, PyArray_DescrFromType(NPY_BOOL),
+                npy_uncertain, 1)
+  REGISTER_CAST(Uncertain_t, npy_bool, &npyuncertain_descr, NPY_BOOL, 0)
+
+#define REGISTER_UFUNC(name, ufuncmodule, ...)                                 \
+  {                                                                            \
+    PyUFuncObject *ufunc =                                                     \
+        (PyUFuncObject *)PyObject_GetAttrString(ufuncmodule, #name);           \
+    if (!ufunc) {                                                              \
+      return -1;                                                               \
+    }                                                                          \
+    int arg_types[] = __VA_ARGS__;                                             \
+    if (sizeof(arg_types) / sizeof(int) != ufunc->nargs) {                     \
+      Py_DECREF(module);                                                       \
+      return -1;                                                               \
+    }                                                                          \
+    if (PyUFunc_RegisterLoopForType((PyUFuncObject *)ufunc, npy_uncertain,     \
+                                    uncertain_ufunc_##name, arg_types,         \
+                                    0) < 0) {                                  \
+      Py_DECREF(module);                                                       \
+      return -1;                                                               \
+    }                                                                          \
+    Py_DECREF(ufunc);                                                          \
+  }
+
+#define REGISTER_NUMPY_UFUNC(name, ...) REGISTER_UFUNC(name, numpy, __VA_ARGS__)
+
+#define REGISTER_NUMPY_UFUNC_BINOP_UNCERTAIN(name)                             \
+  REGISTER_NUMPY_UFUNC(name, {npy_uncertain, npy_uncertain, npy_uncertain})
+#define REGISTER_NUMPY_UFUNC_BINOP_BOOL(name)                                  \
+  REGISTER_NUMPY_UFUNC(name, {npy_uncertain, npy_uncertain, NPY_BOOL})
+
+  REGISTER_NUMPY_UFUNC_BINOP_UNCERTAIN(add)
+  REGISTER_NUMPY_UFUNC_BINOP_UNCERTAIN(subtract)
+  REGISTER_NUMPY_UFUNC_BINOP_UNCERTAIN(multiply)
+  REGISTER_NUMPY_UFUNC_BINOP_UNCERTAIN(divide)
+
+  REGISTER_NUMPY_UFUNC_BINOP_BOOL(equal)
+  REGISTER_NUMPY_UFUNC_BINOP_BOOL(not_equal)
+
+#define NEW_UFUNC(name, doc)                                                   \
+  {                                                                            \
+    PyObject *ufunc = PyUFunc_FromFuncAndData(0, 0, 0, 0, 1, 1, PyUFunc_None,  \
+                                              (char *)#name, (char *)doc, 0);  \
+    if (!ufunc) {                                                              \
+      Py_DECREF(module);                                                       \
+      return -1;                                                               \
+    }                                                                          \
+    PyModule_AddObject(module, #name, (PyObject *)ufunc);                      \
+  }
+
+  NEW_UFUNC(nominal, "uncertain number nominal value")
+  NEW_UFUNC(uncertainty, "uncertain number uncertainity")
+
+#define REGISTER_MODULE_UFUNC(name, ...)                                       \
+  REGISTER_UFUNC(name, module, __VA_ARGS__)
+#define REGISTER_MODULE_UFUNC_UNARY_FLOAT64(name)                              \
+  REGISTER_MODULE_UFUNC(name, {npy_uncertain, NPY_FLOAT64})
+
+  REGISTER_MODULE_UFUNC_UNARY_FLOAT64(nominal)
+  REGISTER_MODULE_UFUNC_UNARY_FLOAT64(uncertainty)
+}
